+++ packages/core/src/components/ProgressBars/LinearProgressBar/__stories__/linearProgressBar.stories.tsx
@@ -1,4 +1,4 @@
-import LinearProgressBar from "../LinearProgressBar";
+import LinearProgressBar, { LinearProgressBarProps } from "../LinearProgressBar";
 import { useMemo } from "react";
 import { createStoryMetaSettingsDecorator } from "../../../../storybook/functions/createStoryMetaSettingsDecorator";
 import colors from "../../../../constants/colors.json";
@@ -20,7 +20,7 @@ export default {
   decorators: metaSettings.decorators
 };
 
-const lineProgressBarTemplate = args => {
+const lineProgressBarTemplate = (args: LinearProgressBarProps) => {
   return <LinearProgressBar className="linear-progress-bar_small-wrapper" {...args} />;
 };
 
@@ -134,7 +134,7 @@ export const ProgressBarAsLoadingIndicator = {
         <div className="linear-progress-bar_img" />
         <div className="linear-progress-bar_aside">
           <b>Frame 697.pg</b>
-          <BreadcrumbsBar className="linear-progress-bar_breadcrumbs">
+          <BreadcrumbsBar type={BreadcrumbsBar.types.INDICATION} className="linear-progress-bar_breadcrumbs">
             <BreadcrumbItem text="Hadas Test" />
             <BreadcrumbItem text="Activity 6" />
           </BreadcrumbsBar>