+++ packages/core/src/components/Tipseen/TipseenTitle.tsx
@@ -3,8 +3,10 @@ import { VibeComponentProps } from "../../types";
 import { getTestId } from "../../tests/test-ids-utils";
 import { ComponentDefaultTestId } from "../../tests/constants";
 import Text from "../Text/Text";
+import cx from "classnames";
+import styles from "./TipseenTitle.module.scss";
 
-interface TipseenTitleProps extends VibeComponentProps {
+export interface TipseenTitleProps extends VibeComponentProps {
   text?: string;
 }
 
@@ -17,7 +19,7 @@ const TipseenTitle: FC<TipseenTitleProps> = ({ text, className, id, "data-testid
       color={Text.colors.ON_PRIMARY}
       aria-level={3}
       maxLines={2}
-      className={className}
+      className={cx(styles.title, className)}
       id={id}
       data-testid={dataTestId || getTestId(ComponentDefaultTestId.TIPSEEN_TITLE, id)}
     >