+++ packages/core/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.92.9](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.92.8...monday-ui-react-core@2.92.9) (2024-02-19)
+
+
+### Bug Fixes
+
+* **EditableTypography:** react to value prop change ([#1971](https://github.com/mondaycom/monday-ui-react-core/issues/1971)) ([de17685](https://github.com/mondaycom/monday-ui-react-core/commit/de176855d4d6578549012225541de9c4d12015e4))
+* **Tipseen:** add gap between title and close button ([#1972](https://github.com/mondaycom/monday-ui-react-core/issues/1972)) ([b0ceea2](https://github.com/mondaycom/monday-ui-react-core/commit/b0ceea2a52ef53d267ad68ba0e8104bfa4186388))
+
+
+
+
+
+## [2.92.8](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.92.7...monday-ui-react-core@2.92.8) (2024-02-18)
+
+
+### Bug Fixes
+
+* **AlertBanner:** fix children type ([#1968](https://github.com/mondaycom/monday-ui-react-core/issues/1968)) ([cf4845d](https://github.com/mondaycom/monday-ui-react-core/commit/cf4845d00e6caf85cfba45b6dd4823fdd9c19452))
+
+
+
+
+
 ## [2.92.7](https://github.com/mondaycom/monday-ui-react-core/compare/monday-ui-react-core@2.92.6...monday-ui-react-core@2.92.7) (2024-02-14)